mindworld/requirements.txt
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
asyncio-dgram==2.2.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
behave==1.2.6
behave-html-pretty-formatter==1.11.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.3
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
dnspython==2.6.1
dominate==2.9.0
EbookLib==0.18
exceptiongroup==1.2.1
Flask==3.0.3
fpdf==1.7.2
google-auth==2.29.0
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itsdangerous==2.2.0
Jinja2==3.1.4
kubernetes==30.1.0
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
mcstatus==11.1.1
msgpack==1.0.8
oauthlib==3.2.2
parse==1.20.1
parse-type==0.6.2
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==306
PyYAML==6.0.1
redis==5.0.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
windows-curses==2.3.3
zope.interface==6.4.post2
